requirements.txt
tensorflow-cpu
streamlit
numpy
scikit-image